requirements.txt
Flask==2.0.1
folium==0.12.1
geopandas==0.14.3
branca==0.7.0
Werkzeug==3.0.1
rasterio==1.3.8
numpy==1.24.3
gunicorn==20.1.0
Pillow==10.0.0 